Update arch-installer.sh
Added support for btrfs

+++ arch-installer.sh
@@ -33,12 +33,33 @@ for i in "${partitions[@]}"; do
 		mkswap ${array[0]}
 		;;
 	"8300")
-		mkfs.ext4 -F ${array[0]}
+ 		[[ filesystem = "brtfs" ]] && mkfs.btrfs ${array[0]}
+		[[ filesystem = "ext4" ]] && mkfs.ext4 -F ${array[0]}
 		;;
 	esac
 done
 
-# Mount partitions
+# Mount Btrfs
+if [[ filesystem = "brtfs" ]]; then
+    mount $(findroot) /mnt
+    for i in "${subvolumes[@]}"; do
+        IFS='|' read -ra array <<< "$i"
+        btrfs subvolume create /mnt/${array[0]}
+    done
+    umount /mnt
+    
+    for i in "${subvolumes[@]}"; do
+        IFS='|' read -ra array <<< "$i"
+        mount -o noatime,compress=zstd,subvol=${array[0]} $(findroot) ${array[1]} --mkdir
+    done
+    
+    IFS='|' read -ra esp <<< ${partitions[0]}
+    mount -o umask=0077 ${esp[0]} ${esp[4]} --mkdir
+    IFS='|' read -ra swap <<< ${partitions[1]}
+    swapon ${esp[0]}
+fi
+
+# Mount Ext4
 mount $(findroot) /mnt
 for i in "${partitions[@]}"; do
 	IFS='|' read -ra array <<<"$i"
